+++ transform-demo/src/main/java/com/robot/transform/demo/bean/Student.java
@@ -21,4 +21,5 @@ public class Student {
     private Integer hobby;
     // 班级-需班级表翻译（class表根据id查询name）
     private Long classId;
+    private String email;
 }
