+++ transform-demo/src/main/java/com/robot/transform/demo/controller/transform/TransformClass.java
@@ -1,9 +1,6 @@
 package com.robot.transform.demo.controller.transform;
 
-import com.fasterxml.jackson.annotation.JacksonAnnotationsInside;
-import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 import com.robot.transform.annotation.Transform;
-import com.robot.transform.serialize.TransformSerializer;
 import org.springframework.core.annotation.AliasFor;
 
 import java.lang.annotation.*;
@@ -17,8 +14,6 @@ import java.lang.annotation.*;
 @Documented
 @Target({ElementType.FIELD})
 @Transform(transformer = ClassTransformer.class)
-@JsonSerialize(using = TransformSerializer.class)
-@JacksonAnnotationsInside
 public @interface TransformClass {
 
     /**
