+++ transform-extension-for-mybatis-flex/src/main/java/com/robot/transform/extend/TransformExtendForMyBatisFlexAutoConfiguration.java
@@ -1,7 +1,6 @@
 package com.robot.transform.extend;
 
 import com.robot.transform.extend.transformer.ForeignKeyTransformer;
-import com.robot.transform.extend.unwrapper.PageUnWrapper;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
@@ -12,14 +11,6 @@ import org.springframework.context.annotation.Configuration;
  */
 @Configuration
 public class TransformExtendForMyBatisFlexAutoConfiguration {
-
-    /**
-     * 注册IPage解包器
-     */
-    @Bean
-    public PageUnWrapper<Object> iPageUnWrapper() {
-        return new PageUnWrapper<>();
-    }
 
     /**
      * 注册外键转换器
