+++ transform-extension-for-mybatis-plus/src/main/java/com/robot/transform/extend/annotation/TransformFK.java
@@ -2,11 +2,8 @@ package com.robot.transform.extend.annotation;
 
 
 import com.baomidou.mybatisplus.core.mapper.BaseMapper;
-import com.fasterxml.jackson.annotation.JacksonAnnotationsInside;
-import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 import com.robot.transform.annotation.Transform;
 import com.robot.transform.extend.transformer.ForeignKeyTransformer;
-import com.robot.transform.serialize.TransformSerializer;
 import org.springframework.core.annotation.AliasFor;
 
 import java.lang.annotation.*;
@@ -21,8 +18,6 @@ import java.lang.annotation.*;
 @Documented
 @Target({ElementType.FIELD})
 @Transform(transformer = ForeignKeyTransformer.class)
-@JsonSerialize(using = TransformSerializer.class)
-@JacksonAnnotationsInside
 public @interface TransformFK {
     /**
      * 目标字段
