+++ transform-spring-boot-starter/pom.xml
@@ -55,11 +55,14 @@
         </dependency>
         <dependency>
             <groupId>io.github.luo-zhan</groupId>
-            <artifactId>easy-enum</artifactId>
+            <artifactId>easy-enum-for-spring</artifactId>
             <version>1.3.1-RELEASE</version>
-            <optional>true</optional>
         </dependency>
-
+        <dependency>
+            <groupId>org.springframework.cloud</groupId>
+            <artifactId>spring-cloud-openfeign-core</artifactId>
+            <version>3.0.4</version>
+        </dependency>
 
     </dependencies>
 </project>
